Feat: adds method to get a list of variated abstract tasks
This method could be useful to be use with the UniqueTaskPipelineValidator.
So far it seems like we always put every abstract task into the
UniqueTaskPipelineValidator because those should be unique within a single
pipeline run. This method uses information from subtypes and the repository
to automatically get all abstract tasks that have been really variated while
implementing the pipeline.

*Caution, not really tested*